fix/ Projeto Pedido configurando nota fiscal inacabado

+++ src/main/java/com/example/pedidos/model/ItemPedido.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class ItemPedido {
-    private int quantidade;
-    private Long produtoId;
+    private long quantidade;
+    private long produtoId;
 }
